requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
confection==0.1.5
cymem==2.0.11
Deprecated==1.2.18
deptry==0.23.0
diskcache==5.6.3
epub2pdf==0.1.0
eval_type_backport==0.2.2
gitdb==4.0.12
GitPython==3.1.44
google==3.0.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
img2pdf==0.6.1
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langcodes==3.5.0
language_data==1.3.0
loguru==0.6.0
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mistralai==1.7.0
mobi==0.3.3
murmurhash==1.0.12
narwhals==1.37.1
numpy==2.2.5
oauthlib==3.2.2
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pikepdf==9.7.0
pillow==11.2.1
preshed==3.0.9
proto-plus==1.26.1
protobuf==6.30.2
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
RapidFuzz==3.13.0
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
requirements-parser==0.11.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
scipy==1.15.2
setuptools==80.1.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
streamlit==1.45.0
tenacity==9.1.2
thinc==8.3.6
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typer==0.15.3
types-setuptools==80.0.0.20250429
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
wrapt==1.17.2
supabase==2.15.2
psycopg2-binary==2.9.9
